requirements.txt
requests>=2.31.0
genanki>=0.13.0
python-dotenv>=1.0.0
google-generativeai
Pillow

